Added dropdown menu for user and removed route to user page.
