+++ movielist-react/src/App.css
@@ -30,6 +30,7 @@ body {
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   padding: 0.5rem;
   justify-content: space-between;
+  align-items: center;
 }
 .header-container .header-container-inner {
   display: flex;
@@ -63,29 +64,62 @@ body {
   color: rgb(35, 90, 92);
 }
 
-.user-container {
+.dropdown-login-btn {
+  padding: 0.5rem;
+  background-color: #35cee4;
+  border: none;
+  border-radius: 5px;
+  color: black;
+  display: inline-block;
+  line-height: 1;
+}
+.dropdown-login-btn:hover {
+  background-color: #2ba8bc;
+}
+
+.dropdown-container {
+  position: relative;
+}
+.dropdown-container .dropdown-container-active {
+  background: none;
+  border: none;
+  padding: 0;
+}
+.dropdown-container .dropdown-container-active:hover {
+  opacity: 0.8;
+}
+.dropdown-container .dropdown-content {
+  position: absolute;
+  right: 0;
+  background: rgb(231, 238, 238);
+  border-radius: 8px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  min-width: 250px;
+}
+.dropdown-container .dropdown-content .user-container {
   padding: 1rem;
 }
-.user-container form {
+.dropdown-container .dropdown-content .user-container form {
   display: flex;
   flex-direction: column;
   gap: 1rem;
 }
-.user-container form input {
+.dropdown-container .dropdown-content .user-container form input {
   padding: 0.5rem;
   border: 1px solid #d1cfcf;
   border-radius: 4px;
 }
-.user-container form button {
+.dropdown-container .dropdown-content .user-container form button {
   padding: 0.5rem;
   background-color: #35cee4;
   border: none;
   border-radius: 4px;
+  color: black;
 }
-.user-container form button:hover {
+.dropdown-container .dropdown-content .user-container form button:hover {
   background-color: #2ba8bc;
 }
-.user-container .userLoginError {
+.dropdown-container .dropdown-content .user-container .userLoginError {
   color: red;
   margin-bottom: 1rem;
 }
