+++ movielist-react/src/App.jsx
@@ -13,34 +13,11 @@ import LiveCommunityChat from "./components/LiveCommunityChat.jsx";
 import Movie from "./components/Movie.jsx";
 import Reviews from "./components/Reviews.jsx";
 import Footer from "./components/Footer.jsx";
-import { MovieProvider, MovieContext } from "./context/MovieContext.jsx";
-import User from "./components/User.jsx";
+import { MovieProvider } from "./context/MovieContext.jsx";
 // @ts-ignore
 import MLlogo from "./assets/MLlogo.png";
-import { IoPersonSharp } from "react-icons/io5";
+import UserMenu from "./components/UserMenu.jsx";
 
-const MovieHeader = () => {
-  const { isLoggedIn, user } = useContext(MovieContext);
-
-  let selectAvatar = "avatar " + "avatar-letters";
-
-  return (
-    <div>
-        <div>
-          <NavLink to="/user">
-            {isLoggedIn ? (
-              <span className={selectAvatar}>
-                <IoPersonSharp />
-                {user.username.charAt(0).toUpperCase()}
-              </span>
-            ) : (
-              <button>Logga in</button>
-            )}
-          </NavLink>
-        </div>
-    </div>
-  );
-};
 
 function App() {
   const [selectedMovie, setSelectedMovie] = useState(null);
@@ -60,7 +37,7 @@ function App() {
             />
             <h2>MovieList</h2>
             </div>
-            <MovieHeader />
+            <UserMenu />
             </header>
           <nav className="menu-container">
             <ul className="menu-list">
@@ -165,7 +142,6 @@ function App() {
               />
               <Route path="/movie/" element={<Movie />} />
               <Route path="/recensioner" element={<Reviews />} />
-              <Route path="/user" element={<User />} />
             </Routes>
           </div>
           <Footer />
