+++ movielist-react/src/App.scss
@@ -42,13 +42,14 @@ body {
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   padding: 0.5rem;
   justify-content: space-between;
+  align-items: center;
 
   .header-container-inner {
     display: flex;
     gap: 0.5rem;
     align-items: center;
 
-    .logoImage{
+    .logoImage {
       height: 60px;
       object-fit: cover;
     }
@@ -57,20 +58,6 @@ body {
       font-size: 30px;
     }
   }
-
-  
-}
-
-.avatar {
-  height: 3rem;
-  width: 3rem;
-  border-radius: 50%;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: .1rem solid rgb(35, 90, 92);
-  color: rgb(35, 90, 92);
 }
 
 // .avatar-letters {
@@ -81,41 +68,91 @@ body {
 //   font-size: 0.8rem;
 // }
 
+.avatar {
+  height: 3rem;
+  width: 3rem;
+  border-radius: 50%;
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 0.1rem solid rgb(35, 90, 92);
+  color: rgb(35, 90, 92);
+}
+
 .avatar > svg {
   height: 50%;
   width: 50%;
   color: rgb(35, 90, 92);
 }
 
-.user-container {
-  padding: 1rem;
-  
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    
-    input {
-      padding: 0.5rem;
-      border: 1px solid #d1cfcf;
-      border-radius: 4px;
-    }
-    
-    button {
-      padding: 0.5rem;
-      background-color: #35cee4;
-      border: none;
-      border-radius: 4px;
-      
-      &:hover {
-        background-color: #2ba8bc;
-      }
+.dropdown-login-btn {
+  padding: 0.5rem;
+  background-color: #35cee4;
+  border: none;
+  border-radius: 5px;
+  color: black;
+  display: inline-block;
+  line-height: 1;
+
+  &:hover {
+    background-color: #2ba8bc;
+  }
+}
+
+.dropdown-container {
+  position: relative;
+
+  .dropdown-container-active {
+    background: none;
+    border: none;
+    padding: 0;
+
+    &:hover {
+      opacity: 0.8;
     }
   }
-  
-  .userLoginError {
-    color: red;
-    margin-bottom: 1rem;
+
+  .dropdown-content {
+    position: absolute;
+    right: 0;
+    background: rgb(231, 238, 238);
+    border-radius: 8px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+    min-width: 250px;
+
+    .user-container {
+      padding: 1rem;
+
+      form {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+
+        input {
+          padding: 0.5rem;
+          border: 1px solid #d1cfcf;
+          border-radius: 4px;
+        }
+
+        button {
+          padding: 0.5rem;
+          background-color: #35cee4;
+          border: none;
+          border-radius: 4px;
+          color: black;
+
+          &:hover {
+            background-color: #2ba8bc;
+          }
+        }
+      }
+
+      .userLoginError {
+        color: red;
+        margin-bottom: 1rem;
+      }
+    }
   }
 }
